+++ 01-kms-core/04-kms-starter-ui/src/main/java/eu/isygoit/ui/kms/views/cryptography/random/dialog/DeleteRandomKeyDialog.java
@@ -20,8 +20,14 @@ public class DeleteRandomKeyDialog extends PinBaseActionDialog {
 
     @Override
     protected boolean onOk() {
+        if (!validatePin()) {
+            append("Invalid confirmation code");
+            return false;
+        }
+
         try {
             keyService.deleteRandomKey(keyName);
+            append("Key deleted");
             return true;
         } catch (Exception e) {
             append("Delete failed: " + e.getMessage());
